requirements.txt
PySimpleGUI==4.60.5
paramiko==3.3.1
pycryptodome==3.18.0